Operations added
Now when the required operation is typed and the equal button is
pressed, the operation is evaluated and the result is shown in the
label. But there is an error. Once the equal button is pressed, clear
button must be used. If any other button is pressed an error is shown.